smartbudget/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
pandas>=2.0
scikit-learn>=1.2
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
